Rename FormExportTable to FormsTable
