+++ src/org/opendatakit/briefcase/ui/export/components/FormsTable.java
@@ -14,14 +14,14 @@
  * the License.
  */
 
-package org.opendatakit.briefcase.ui.export;
+package org.opendatakit.briefcase.ui.export.components;
 
 import static javax.swing.SortOrder.ASCENDING;
-import static org.opendatakit.briefcase.ui.export.FormExportTableModel.DETAIL_BUTTON_COL;
-import static org.opendatakit.briefcase.ui.export.FormExportTableModel.FORM_NAME_COL;
-import static org.opendatakit.briefcase.ui.export.FormExportTableModel.HEADERS;
-import static org.opendatakit.briefcase.ui.export.FormExportTableModel.SELECTED_CHECKBOX_COL;
-import static org.opendatakit.briefcase.ui.export.FormExportTableModel.buildDetailButton;
+import static org.opendatakit.briefcase.ui.export.components.FormsTableViewModel.DETAIL_BUTTON_COL;
+import static org.opendatakit.briefcase.ui.export.components.FormsTableViewModel.FORM_NAME_COL;
+import static org.opendatakit.briefcase.ui.export.components.FormsTableViewModel.HEADERS;
+import static org.opendatakit.briefcase.ui.export.components.FormsTableViewModel.SELECTED_CHECKBOX_COL;
+import static org.opendatakit.briefcase.ui.export.components.FormsTableViewModel.buildDetailButton;
 
 import java.awt.Dimension;
 import java.awt.event.MouseEvent;
@@ -37,10 +37,10 @@ import javax.swing.table.TableRowSorter;
 import org.bushe.swing.event.annotation.AnnotationProcessor;
 import org.opendatakit.briefcase.ui.reused.MouseListenerBuilder;
 
-class FormExportTable extends JTable {
+public class FormsTable extends JTable {
   private static final long serialVersionUID = 8511088963758308085L;
 
-  FormExportTable(FormExportTableModel tableModel) {
+  public FormsTable(FormsTableViewModel tableModel) {
     super(tableModel);
     AnnotationProcessor.process(this);
 
@@ -65,7 +65,7 @@ class FormExportTable extends JTable {
 
     setFillsViewportHeight(true);
 
-    TableRowSorter<FormExportTableModel> sorter = sortBy((FormExportTableModel) this.dataModel, FORM_NAME_COL, ASCENDING);
+    TableRowSorter<FormsTableViewModel> sorter = sortBy((FormsTableViewModel) this.dataModel, FORM_NAME_COL, ASCENDING);
     setRowSorter(sorter);
     sorter.sort();
   }
